Add HTTP code handling to THttpClient (#380)
Before this change Impyla is not checking HTTP return codes when
using the hs2-http protocol. Impyla is sending a request message
(e.g. send_CloseOperation) but the HTTP call to send this message may
fail. This will result in a failure when reading the reply (e.g. in
recv_CloseOperation) as there is no reply data to read. This will
typically result in an 'EOFError'.

In code that overrides THttpClient.flush(), check the HTTP code that is
returned after the HTTP call is made. If the code is not 1XX
(informational response) or 2XX (successful) then throw an RPCException.

This change does not contain any attempt to recover from an HTTP failures
but it does allow the failure to be detected and a message to be
printed.

In future it may be possible to retry after certain HTTP errors.

Testing:
- Add a new test that tries to connect to an HTTP
  server that always returns a 503 error. Check that an appropriate
  Exception(FIXME) is thrown