Add HTTP code handling to THttpClient (#380)
Before this change Impyla is not checking HTTP return codes when
using the hs2-http protocol. Impyla is sending a request message
(e.g. send_CloseOperation) but the HTTP call to send this message may
fail. This will result in a failure when reading the reply (e.g. in
recv_CloseOperation) as there is no reply data to read. This will
typically result in an 'EOFError'.

Replace use of HTTPClient with ImpalaHttpClient. This code is copied from
Impala Shell. Make some small changes to enable this code to run on
Python3 as well as Python2.

This change does not contain any attempt to recover from an HTTP failures
but it does allow the failure to be detected and a message to be
printed.

In future it may be possible to retry after certain HTTP errors.

Testing:
- Add a new test that tries to connect to an HTTP
  server that always returns a 503 error. Check that an RPCError
  is thrown conaining a description of the error.